[backend] improvements - add intermediate record to sort as stream - sort raw exercises with stream - fix percentage type - add tracing in scenario statistics route - add tests for global scores route and statistics service - truncate global score success percentage to have only one decimal - simplify scenario statistics obtention

+++ openbas-api/src/main/java/io/openbas/rest/scenario/response/GlobalScoreBySimulationEndDate.java
@@ -6,5 +6,4 @@ import java.time.Instant;
 
 public record GlobalScoreBySimulationEndDate(
     @JsonProperty("simulation_end_date") @NotNull Instant simulationEndDate,
-    @JsonProperty("global_score_success_percentage") @NotNull
-        double globalScoreSuccessPercentage) {}
+    @JsonProperty("global_score_success_percentage") @NotNull float globalScoreSuccessPercentage) {}
